test(org): split boot.test.ts (1444 lines) into per-concern files
Last of the three oversized files (after #421/Request, #422/ax). boot's
seven helpers were interleaved through the file (bootstrap, runGate,
escapeRegex, bootstrapWithMembers, registerMember, makeRequestWithTarget,
makeRequestSessioned, plus the GATE path); centralized into
_bootHelpers.ts, then the tests split by concern:
  - boot.test.ts (trimmed) — JSON-shape stability, actor/role,
    misconfigured-cwd, content_root_health, content-root disclosure, tail
  - bootReviewedAuthored.test.ts — the reviewed-authored surface
  - bootOverlap.test.ts — active_overlapping_targets + same-actor
    parallel-session detection (#234 / #249)
  - bootWarnings.test.ts — C3 silent-fallback warnings

Same 42 tests (verified by count + a focused run of all four files, then
a clean-dist full run to rule out trap_dist_stale_after_branch_switch
double-counting). No behavior change. The three big test files
(Request 1158, ax 1549, boot 1444) are now all decomposed.

Full suite (clean dist) 1790 pass / 0 fail.
