+++ tests/interface/boot.test.ts
@@ -13,34 +13,16 @@ import { tmpdir } from 'node:os';
 import { join, resolve, dirname } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { makeTempRoot } from '../util/tempRoot.js';
-
-// At runtime this file lives under dist/tests/interface/, so we walk
-// three levels up (interface → tests → dist → repo root) to reach bin/.
-const here = dirname(fileURLToPath(import.meta.url));
-const GATE = resolve(here, '../../../bin/gate.mjs');
-
-function bootstrap(): { root: string; cleanup: () => void } {
-  const root = makeTempRoot('guild-boot-');
-  writeFileSync(
-    join(root, 'guild.config.yaml'),
-    'content_root: .\nhost_names: [human]\n',
-  );
-  mkdirSync(join(root, 'members'));
-  writeFileSync(
-    join(root, 'members', 'alice.yaml'),
-    'name: alice\ncategory: professional\nactive: true\n',
-  );
-  return { root, cleanup: () => rmSync(root, { recursive: true, force: true }) };
-}
-
-function runGate(cwd: string, args: string[], env: Record<string, string> = {}): { stdout: string; status: number } {
-  const result = spawnSync(process.execPath, [GATE, ...args], {
-    cwd,
-    env: { ...process.env, ...env },
-    encoding: 'utf8',
-  });
-  return { stdout: result.stdout, status: result.status ?? -1 };
-}
+import {
+  GATE,
+  bootstrap,
+  runGate,
+  escapeRegex,
+  bootstrapWithMembers,
+  registerMember,
+  makeRequestWithTarget,
+  makeRequestSessioned,
+} from './_bootHelpers.js';
 
 test('gate boot: JSON top-level keys are stable', () => {
   const { root, cleanup } = bootstrap();
@@ -507,938 +489,3 @@ test('gate boot: --tail <N> overrides the default', () => {
   }
 });
 
-function escapeRegex(s: string): string {
-  return s.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-}
-
-// ---------------------------------------------------------------
-// reviewed-authored surface (designed in 2026-05-06-0001).
-//
-// When peers land reviews on a request the actor authored, boot
-// lifts them via verbs_available_now.actionable[] (verb=show) and
-// suggested_next (when no higher-priority transition is open).
-// Boundary scope is the Request aggregate: status_log / reviews /
-// thanks all advance it. Message/issue writes do NOT advance it.
-//
-// Tests below pin the gap-1 (thanks integration), gap-2 (cap), and
-// the higher-priority-suppression invariant Devil v3 ratify named.
-// ---------------------------------------------------------------
-
-function bootstrapWithMembers(): { root: string; cleanup: () => void } {
-  // Two-member fixture so reviewer != author. alice authors, bob
-  // reviews — the minimal shape the reviewed-authored predicate
-  // needs.
-  const root = mkdtempSync(join(tmpdir(), 'guild-boot-rev-'));
-  writeFileSync(
-    join(root, 'guild.config.yaml'),
-    'content_root: .\nhost_names: [human]\n',
-  );
-  mkdirSync(join(root, 'members'));
-  writeFileSync(
-    join(root, 'members', 'alice.yaml'),
-    'name: alice\ncategory: professional\nactive: true\n',
-  );
-  writeFileSync(
-    join(root, 'members', 'bob.yaml'),
-    'name: bob\ncategory: professional\nactive: true\n',
-  );
-  return { root, cleanup: () => rmSync(root, { recursive: true, force: true }) };
-}
-
-test('gate boot: reviewed-authored surfaces when peer reviews land on authored request', () => {
-  const { root, cleanup } = bootstrapWithMembers();
-  try {
-    // alice authors a fast-track (reaches completed in one shot).
-    const filed = runGate(
-      root,
-      [
-        'fast-track',
-        '--action',
-        'demo work',
-        '--reason',
-        'reviewed-authored pin',
-        '--format',
-        'json',
-      ],
-      { GUILD_ACTOR: 'alice' },
-    );
-    assert.equal(filed.status, 0);
-    const id = JSON.parse(filed.stdout).id;
-    // bob reviews. bob's review is the boundary-crossing event.
-    const reviewed = runGate(
-      root,
-      [
-        'review',
-        id,
-        '--lense',
-        'devil',
-        '--verdict',
-        'ok',
-        '--comment',
-        'lgtm',
-      ],
-      { GUILD_ACTOR: 'bob' },
-    );
-    assert.equal(reviewed.status, 0);
-
-    // alice boots — should see reviewed-authored.
-    const { stdout } = runGate(root, ['boot'], { GUILD_ACTOR: 'alice' });
-    const payload = JSON.parse(stdout);
-    assert.equal(payload.suggested_next?.verb, 'show');
-    assert.equal(payload.suggested_next?.args?.id, id);
-    assert.match(payload.suggested_next?.reason ?? '', /you authored/);
-    assert.match(payload.suggested_next?.reason ?? '', /1 review/);
-    assert.equal(payload.status.reviews_unseen, 1);
-
-    // actionable[] mirrors it.
-    const actionable = payload.verbs_available_now.actionable;
-    assert.ok(
-      actionable.some(
-        (a: { verb: string; id: string }) => a.verb === 'show' && a.id === id,
-      ),
-      'reviewed-authored entry must appear in actionable[]',
-    );
-  } finally {
-    cleanup();
-  }
-});
-
-test('gate boot: reviewed-authored boundary advances when author writes a thank (gap-1)', () => {
-  // Devil v3 concern: addThank does not touch status_log, so a
-  // thanks-only response from the author would not advance the
-  // boundary in v2. Pin that v3's thanks integration prevents the
-  // surface from sticking after the author thanks the reviewer.
-  const { root, cleanup } = bootstrapWithMembers();
-  try {
-    const filed = runGate(
-      root,
-      [
-        'fast-track',
-        '--action',
-        'thanks-advances-boundary',
-        '--reason',
-        'gap-1 regression pin',
-        '--format',
-        'json',
-      ],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const id = JSON.parse(filed.stdout).id;
-    runGate(
-      root,
-      [
-        'review',
-        id,
-        '--lense',
-        'devil',
-        '--verdict',
-        'ok',
-        '--comment',
-        'lgtm',
-      ],
-      { GUILD_ACTOR: 'bob' },
-    );
-    // Confirm surface is up before the thank.
-    const before = JSON.parse(
-      runGate(root, ['boot'], { GUILD_ACTOR: 'alice' }).stdout,
-    );
-    assert.equal(before.suggested_next?.verb, 'show');
-    assert.equal(before.status.reviews_unseen, 1);
-
-    // alice thanks bob. addThank pushes onto thanks[] (no status_log
-    // touch). v3 boundary computes lastAuthoredWriteAt over thanks[]
-    // too, so the surface must clear.
-    const thanked = runGate(
-      root,
-      ['thank', 'bob', '--for', id, '--reason', 'thanks for the review'],
-      { GUILD_ACTOR: 'alice' },
-    );
-    assert.equal(thanked.status, 0);
-
-    const after = JSON.parse(
-      runGate(root, ['boot'], { GUILD_ACTOR: 'alice' }).stdout,
-    );
-    // suggested_next must no longer point at this id (no other open
-    // loops in this fixture, so it should be null).
-    assert.equal(
-      after.suggested_next,
-      null,
-      'thank by author must advance boundary so reviewed-authored clears',
-    );
-    assert.ok(
-      !('reviews_unseen' in after.status) || after.status.reviews_unseen === 0,
-      'reviews_unseen must be cleared (or absent) after thank',
-    );
-    const actionable = after.verbs_available_now.actionable as Array<{
-      verb: string;
-      id: string;
-    }>;
-    assert.ok(
-      !actionable.some((a) => a.verb === 'show' && a.id === id),
-      'actionable[] must drop the reviewed-authored entry after thank',
-    );
-  } finally {
-    cleanup();
-  }
-});
-
-test('gate boot: reviewed-authored actionable[] is capped at 5 entries', () => {
-  // Voice budget: an actor with N authored × M reviews could balloon
-  // the boot payload — boot is on the hot path. Cap actionable[] to
-  // 5; the running total still surfaces via status.reviews_unseen.
-  const { root, cleanup } = bootstrapWithMembers();
-  try {
-    const ids: string[] = [];
-    // 7 authored requests by alice — file ALL of them first, THEN
-    // bob reviews each. If we interleaved (fast-track then review,
-    // repeat), each subsequent fast-track would advance alice's
-    // boundary (status_log write) past the prior review, and only
-    // the most recent review would remain "unseen". Batching keeps
-    // every review strictly after alice's last write, which is the
-    // shape the cap is protecting against.
-    for (let i = 0; i < 7; i++) {
-      const filed = runGate(
-        root,
-        [
-          'fast-track',
-          '--action',
-          `cap probe ${i}`,
-          '--reason',
-          'cap-5 pin',
-          '--format',
-          'json',
-        ],
-        { GUILD_ACTOR: 'alice' },
-      );
-      assert.equal(filed.status, 0);
-      ids.push(JSON.parse(filed.stdout).id);
-    }
-    for (const id of ids) {
-      const reviewed = runGate(
-        root,
-        [
-          'review',
-          id,
-          '--lense',
-          'devil',
-          '--verdict',
-          'ok',
-          '--comment',
-          'lgtm',
-        ],
-        { GUILD_ACTOR: 'bob' },
-      );
-      assert.equal(reviewed.status, 0);
-    }
-
-    const { stdout } = runGate(root, ['boot'], { GUILD_ACTOR: 'alice' });
-    const payload = JSON.parse(stdout);
-    const showEntries = (
-      payload.verbs_available_now.actionable as Array<{ verb: string }>
-    ).filter((a) => a.verb === 'show');
-    assert.equal(
-      showEntries.length,
-      5,
-      'reviewed-authored entries in actionable[] must be capped at 5',
-    );
-    // Running total must reflect the full 7, not the capped view.
-    assert.equal(payload.status.reviews_unseen, 7);
-  } finally {
-    cleanup();
-  }
-});
-
-test('gate boot: reviewed-authored is suppressed when a higher-priority transition is open', () => {
-  // Devil v3 ratify invariant: reviewed-authored sits at PRIORITY=4,
-  // appended only when the four state-transition kinds are empty.
-  // If alice has an executing-mine alongside a reviewed authored
-  // request, suggested_next must point at complete, not show.
-  const { root, cleanup } = bootstrapWithMembers();
-  try {
-    // Path A: authored request that gets a peer review (no exec to alice).
-    const filedA = runGate(
-      root,
-      [
-        'fast-track',
-        '--action',
-        'reviewed (low priority)',
-        '--reason',
-        'priority pin',
-        '--format',
-        'json',
-      ],
-      { GUILD_ACTOR: 'alice' },
-    );
-    const idA = JSON.parse(filedA.stdout).id;
-    runGate(
-      root,
-      [
-        'review',
-        idA,
-        '--lense',
-        'devil',
-        '--verdict',
-        'ok',
-        '--comment',
-        'lgtm',
-      ],
-      { GUILD_ACTOR: 'bob' },
-    );
-
-    // Path B: a request that ends up executing-by-alice. Use the
-    // four-step lifecycle so we can stop at executing.
-    const filedB = runGate(
-      root,
-      [
-        'request',
-        '--action',
-        'executing (high priority)',
-        '--reason',
-        'priority pin',
-        '--executors',
-        'alice',
-        '--format',
-        'json',
-      ],
-      { GUILD_ACTOR: 'alice' },
-    );
-    assert.equal(filedB.status, 0);
-    const idB = JSON.parse(filedB.stdout).id;
-    // Approve as host, then execute as alice.
-    const approved = runGate(root, ['approve', idB], { GUILD_ACTOR: 'human' });
-    assert.equal(approved.status, 0);
-    const executed = runGate(root, ['execute', idB], { GUILD_ACTOR: 'alice' });
-    assert.equal(executed.status, 0);
-
-    const { stdout } = runGate(root, ['boot'], { GUILD_ACTOR: 'alice' });
-    const payload = JSON.parse(stdout);
-    // Highest-priority transition wins suggested_next.
-    assert.equal(payload.suggested_next?.verb, 'complete');
-    assert.equal(payload.suggested_next?.args?.id, idB);
-    // reviewed-authored must NOT contaminate actionable[] when
-    // transitions are present (the if-guard is `out.length === 0`).
-    const actionable = payload.verbs_available_now.actionable as Array<{
-      verb: string;
-      id: string;
-    }>;
-    assert.ok(
-      !actionable.some((a) => a.verb === 'show' && a.id === idA),
-      'reviewed-authored must be suppressed when state-transition work is pending',
-    );
-  } finally {
-    cleanup();
-  }
-});
-
-test('gate boot: host actor does NOT emit the "hosts have no inbox" warning', () => {
-  // Pre-fix, every boot run by a host actor produced a 7-line warning
-  // block ("inbox enrichment failed for actor=eris ... hosts do not
-  // have inboxes"). That fact is already conveyed by `role: 'host'`
-  // in the payload; emitting it as a warning every session inverts
-  // principle 09 (orientation-disclosure: surface surprising cases,
-  // stay quiet otherwise — being a host is not surprising state).
-  // The fix skips inbox enrichment for role='host' so the warning
-  // never fires on the host path.
-  const { root, cleanup } = bootstrap(); // host_names: [human]
-  try {
-    const { stdout, status } = runGate(root, ['boot'], {
-      GUILD_ACTOR: 'human',
-    });
-    assert.equal(status, 0);
-    const payload = JSON.parse(stdout);
-    assert.equal(payload.role, 'host', 'host bootstrap precondition');
-    const inboxWarning = (payload.warnings as string[]).find((w) =>
-      w.includes('inbox enrichment failed'),
-    );
-    assert.equal(
-      inboxWarning,
-      undefined,
-      'host boot must NOT raise inbox-enrichment warning (by-design no-inbox)',
-    );
-    // Sanity: a typo'd actor still surfaces the warning (the fix only
-    // suppresses the host path, not unknown-actor diagnostics).
-    const { stdout: typoStdout } = runGate(root, ['boot'], {
-      GUILD_ACTOR: 'nope-typo',
-    });
-    const typoPayload = JSON.parse(typoStdout);
-    assert.equal(typoPayload.role, 'unknown');
-    const typoWarning = (typoPayload.warnings as string[]).find((w) =>
-      w.includes('inbox enrichment failed'),
-    );
-    assert.ok(
-      typoWarning,
-      'unknown actor must still surface inbox-enrichment warning (typo diagnostic)',
-    );
-  } finally {
-    cleanup();
-  }
-});
-
-test('gate boot: executing wave where author≠executor does NOT surface to the author', () => {
-  // Regression for the actionable-misattribution friction observed
-  // 2026-05-13 on req 2026-05-08-0012 (author=eris, executors=[miki]).
-  // boot surfaced `→ next: gate complete <id> --by eris` to the author,
-  // but `gate complete --by <author>` errors with "not in this wave's
-  // executors" — the suggestion is unactionable. The fix tightens the
-  // executing-mine predicate so the author-only path matches only when
-  // the executor list is empty (legacy / self-execute fallback).
-  const { root, cleanup } = bootstrapWithMembers();
-  try {
-    // alice authors a wave executing-by-bob (author ≠ executor).
-    const filed = runGate(
-      root,
-      [
-        'request',
-        '--action',
-        'work for bob',
-        '--reason',
-        'author-not-executor pin',
-        '--executors',
-        'bob',
-        '--format',
-        'json',
-      ],
-      { GUILD_ACTOR: 'alice' },
-    );
-    assert.equal(filed.status, 0);
-    const id = JSON.parse(filed.stdout).id;
-    assert.equal(
-      runGate(root, ['approve', id], { GUILD_ACTOR: 'human' }).status,
-      0,
-    );
-    assert.equal(
-      runGate(root, ['execute', id], { GUILD_ACTOR: 'bob' }).status,
-      0,
-    );
-
-    // alice's boot must NOT name her as `--by` on a complete/fail of
-    // this id — she can't dispatch either.
-    const { stdout } = runGate(root, ['boot'], { GUILD_ACTOR: 'alice' });
-    const payload = JSON.parse(stdout);
-    const actionable = payload.verbs_available_now.actionable as Array<{
-      verb: string;
-      id: string;
-    }>;
-    assert.ok(
-      !actionable.some(
-        (a) => a.id === id && (a.verb === 'complete' || a.verb === 'fail'),
-      ),
-      'author-only must not appear in actionable[] complete/fail when executors list names someone else',
-    );
-    assert.notEqual(
-      payload.suggested_next?.args?.id,
-      id,
-      'author-only must not be steered toward complete/fail on a wave executing-by-someone-else',
-    );
-
-    // bob's boot, by contrast, MUST surface the actionable.
-    const { stdout: bobStdout } = runGate(root, ['boot'], {
-      GUILD_ACTOR: 'bob',
-    });
-    const bobPayload = JSON.parse(bobStdout);
-    const bobActionable = bobPayload.verbs_available_now.actionable as Array<{
-      verb: string;
-      id: string;
-    }>;
-    assert.ok(
-      bobActionable.some((a) => a.id === id && a.verb === 'complete'),
-      'assigned executor must still see complete in actionable[]',
-    );
-    assert.equal(bobPayload.suggested_next?.verb, 'complete');
-    assert.equal(bobPayload.suggested_next?.args?.id, id);
-    assert.equal(bobPayload.suggested_next?.args?.by, 'bob');
-  } finally {
-    cleanup();
-  }
-});
-
-test('computeLastAuthoredWriteAt aggregates across status_log, reviews, and thanks', async () => {
-  // Direct unit test — bypass the CLI to assert the aggregation
-  // independently of the boot wiring. status_log (transitions),
-  // reviews[] (judgements), and thanks[] (appreciation) all
-  // contribute to the boundary. Latest of the three wins.
-  const { computeLastAuthoredWriteAt } = await import(
-    // Sibling-test pattern: source-relative spec, resolves to
-    // dist/src after tsc emit (matches schema.test.ts, voices.test.ts).
-    '../../src/interface/gate/handlers/boot.js'
-  );
-
-  // Stub `Request`-shaped objects with just the getters
-  // computeLastAuthoredWriteAt reads. Plain objects suffice — the
-  // function uses no Request methods, only the three array getters.
-  // Mirror the in-memory shape: status_log carries `by: string` (raw
-  // shape on entries), while reviews[]/thanks[] expose `by: MemberName`
-  // via getters — so the stub uses `{ value }` for those.
-  type MemberNameStub = { value: string };
-  type Stub = {
-    statusLog: ReadonlyArray<{ by: string; at: string }>;
-    reviews: ReadonlyArray<{ by: MemberNameStub; at: string }>;
-    thanks: ReadonlyArray<{ by: MemberNameStub; at: string }>;
-  };
-  const stubs: Stub[] = [
-    {
-      statusLog: [{ by: 'alice', at: '2026-05-01T10:00:00.000Z' }],
-      reviews: [{ by: { value: 'alice' }, at: '2026-05-02T10:00:00.000Z' }],
-      thanks: [{ by: { value: 'alice' }, at: '2026-05-03T10:00:00.000Z' }],
-    },
-    {
-      statusLog: [{ by: 'bob', at: '2026-05-04T10:00:00.000Z' }],
-      reviews: [],
-      thanks: [],
-    },
-  ];
-
-  const lastAlice = computeLastAuthoredWriteAt(
-    'alice',
-    stubs as unknown as Parameters<typeof computeLastAuthoredWriteAt>[1],
-  );
-  // Latest of the three alice writes is the thank at 2026-05-03.
-  assert.equal(lastAlice, '2026-05-03T10:00:00.000Z');
-
-  // bob has no thanks/reviews; only status_log contributes.
-  const lastBob = computeLastAuthoredWriteAt(
-    'bob',
-    stubs as unknown as Parameters<typeof computeLastAuthoredWriteAt>[1],
-  );
-  assert.equal(lastBob, '2026-05-04T10:00:00.000Z');
-
-  // Actor with no writes anywhere yields null.
-  const lastNobody = computeLastAuthoredWriteAt(
-    'nobody',
-    stubs as unknown as Parameters<typeof computeLastAuthoredWriteAt>[1],
-  );
-  assert.equal(lastNobody, null);
-});
-
-// active_overlapping_targets — cross-session race surface (#234).
-//
-// Surfaces active (pending|approved|executing) requests that share
-// the same `target` so a booting agent sees "someone else is on
-// it" before pre-empting. Phase 1: detection + warning, no refuse.
-// Refuse-on-create lives with #227 (swarm profile epic).
-
-function registerMember(root: string, name: string): void {
-  runGate(root, ['register', '--name', name]);
-}
-
-function makeRequestWithTarget(
-  root: string,
-  from: string,
-  action: string,
-  target: string,
-): string {
-  const r = spawnSync(
-    process.execPath,
-    [
-      GATE,
-      'request',
-      '--from', from,
-      '--action', action,
-      '--reason', 'overlap test',
-      '--target', target,
-      '--format', 'json',
-    ],
-    { cwd: root, env: { ...process.env }, encoding: 'utf8' },
-  );
-  if (r.status !== 0) {
-    throw new Error(`gate request failed: ${r.stderr}`);
-  }
-  const j = JSON.parse(r.stdout);
-  return j.id ?? j.request_id;
-}
-
-test('gate boot: active_overlapping_targets surfaces two pending requests on the same target', () => {
-  const { root, cleanup } = bootstrap();
-  try {
-    registerMember(root, 'leysia');
-    const id1 = makeRequestWithTarget(root, 'alice', 'work A', 'data/guild/templates');
-    const id2 = makeRequestWithTarget(root, 'leysia', 'work B', 'data/guild/templates');
-
-    const { stdout } = runGate(root, ['boot']);
-    const payload = JSON.parse(stdout);
-    assert.equal(Array.isArray(payload.active_overlapping_targets), true);
-    assert.equal(payload.active_overlapping_targets.length, 1);
-    const entry = payload.active_overlapping_targets[0];
-    assert.equal(entry.target, 'data/guild/templates');
-    assert.equal(entry.requests.length, 2);
-    // Sorted by id ascending — deterministic across boots.
-    assert.deepEqual(
-      entry.requests.map((r: { id: string }) => r.id),
-      [id1, id2].sort(),
-    );
-    // Each entry carries state + executors[]. claimed_by is omitted
-    // for unclaimed waves (omit-when-undefined convention).
-    for (const r of entry.requests) {
-      assert.ok(['pending', 'approved', 'executing'].includes(r.state));
-      assert.ok(Array.isArray(r.executors));
-      assert.equal('claimed_by' in r, false);
-    }
-  } finally {
-    cleanup();
-  }
-});
-
-test('gate boot: active_overlapping_targets is empty when targets differ', () => {
-  const { root, cleanup } = bootstrap();
-  try {
-    registerMember(root, 'leysia');
-    makeRequestWithTarget(root, 'alice', 'work A', 'src/foo');
-    makeRequestWithTarget(root, 'leysia', 'work B', 'src/bar');
-
-    const { stdout } = runGate(root, ['boot']);
-    const payload = JSON.parse(stdout);
-    assert.deepEqual(payload.active_overlapping_targets, []);
-  } finally {
-    cleanup();
-  }
-});
-
-test('gate boot: active_overlapping_targets ignores requests with no target', () => {
-  // Two requests, both untargeted → no group key → no overlap.
-  // Untargeted overlap is not a coordination signal (the freeform
-  // target is the only handle for "same wave").
-  const { root, cleanup } = bootstrap();
-  try {
-    registerMember(root, 'leysia');
-    const r1 = spawnSync(
-      process.execPath,
-      [GATE, 'request', '--from', 'alice', '--action', 'a', '--reason', 'r', '--format', 'json'],
-      { cwd: root, env: { ...process.env }, encoding: 'utf8' },
-    );
-    assert.equal(r1.status, 0, r1.stderr);
-    const r2 = spawnSync(
-      process.execPath,
-      [GATE, 'request', '--from', 'leysia', '--action', 'b', '--reason', 'r', '--format', 'json'],
-      { cwd: root, env: { ...process.env }, encoding: 'utf8' },
-    );
-    assert.equal(r2.status, 0, r2.stderr);
-
-    const { stdout } = runGate(root, ['boot']);
-    const payload = JSON.parse(stdout);
-    assert.deepEqual(payload.active_overlapping_targets, []);
-  } finally {
-    cleanup();
-  }
-});
-
-test('gate boot: active_overlapping_targets carries claim_held marker for claimed wave', () => {
-  const { root, cleanup } = bootstrap();
-  try {
-    // Overlap text-mode rendering is profile=swarm only (#323).
-    writeFileSync(
-      join(root, 'guild.config.yaml'),
-      'content_root: .\nhost_names: [human]\nprofile: swarm\n',
-    );
-    registerMember(root, 'leysia');
-    // Use --executor to populate the executors[] slot the surface
-    // renders (matching the issue's example output shape, which
-    // names the executor next to the id).
-    const r1 = spawnSync(
-      process.execPath,
-      [
-        GATE, 'request',
-        '--from', 'alice',
-        '--executors', 'alice',
-        '--action', 'work A',
-        '--reason', 'overlap test',
-        '--target', 'shared/path',
-        '--format', 'json',
-      ],
-      { cwd: root, env: { ...process.env }, encoding: 'utf8' },
-    );
-    assert.equal(r1.status, 0, r1.stderr);
-    const id1 = JSON.parse(r1.stdout).id;
-    const r2 = spawnSync(
-      process.execPath,
-      [
-        GATE, 'request',
-        '--from', 'leysia',
-        '--executors', 'leysia',
-        '--action', 'work B',
-        '--reason', 'overlap test',
-        '--target', 'shared/path',
-        '--format', 'json',
-      ],
-      { cwd: root, env: { ...process.env }, encoding: 'utf8' },
-    );
-    assert.equal(r2.status, 0, r2.stderr);
-
-    // Stake an exclusive claim on the first request.
-    const claim = runGate(root, ['claim', id1, '--by', 'alice']);
-    assert.equal(claim.status, 0);
-
-    const { stdout } = runGate(root, ['boot']);
-    const payload = JSON.parse(stdout);
-    const entry = payload.active_overlapping_targets[0];
-    const claimed = entry.requests.find((r: { id: string }) => r.id === id1);
-    assert.equal(claimed.claimed_by, 'alice');
-    // Text mode renders the marker too (`claim_held` flag).
-    const t = runGate(root, ['boot', '--format', 'text']);
-    assert.match(t.stdout, /active waves with overlapping target:/);
-    assert.match(t.stdout, new RegExp(`${id1} \\(alice, pending, claim_held\\)`));
-    assert.match(t.stdout, /target: shared\/path/);
-    assert.match(t.stdout, /coordinate via .gate witness/);
-  } finally {
-    cleanup();
-  }
-});
-
-test('gate boot: active_overlapping_targets text section is omitted when no overlap', () => {
-  // Voice budget — fresh roots / single-wave roots should not see
-  // the warning header line. Empty array is a JSON contract; text
-  // mode silences entirely.
-  const { root, cleanup } = bootstrap();
-  try {
-    const t = runGate(root, ['boot', '--format', 'text']);
-    assert.equal(t.status, 0);
-    assert.doesNotMatch(t.stdout, /active waves with overlapping target/);
-  } finally {
-    cleanup();
-  }
-});
-
-// ---- Slice 4 (#249): same-actor parallel-session detection ---------
-
-function makeRequestSessioned(
-  root: string,
-  from: string,
-  action: string,
-  target: string,
-  sessionId: string,
-): string {
-  const r = spawnSync(
-    process.execPath,
-    [
-      GATE,
-      'request',
-      '--from', from,
-      '--action', action,
-      '--reason', 'overlap test',
-      '--target', target,
-      '--format', 'json',
-    ],
-    {
-      cwd: root,
-      env: { ...process.env, GUILD_SESSION_ID: sessionId },
-      encoding: 'utf8',
-    },
-  );
-  if (r.status !== 0) {
-    throw new Error(`gate request failed: ${r.stderr}`);
-  }
-  return JSON.parse(r.stdout).id as string;
-}
-
-test('gate boot: overlap requests carry opened_by_session when stamped (#249 slice 4)', () => {
-  const { root, cleanup } = bootstrap();
-  try {
-    makeRequestSessioned(root, 'alice', 'work A', 'src/foo', 'alice-tmux-1');
-    makeRequestSessioned(root, 'alice', 'work B', 'src/foo', 'alice-tmux-1');
-
-    const { stdout } = runGate(root, ['boot']);
-    const payload = JSON.parse(stdout);
-    const entry = payload.active_overlapping_targets[0];
-    assert.equal(entry.requests.length, 2);
-    for (const r of entry.requests) {
-      assert.equal(r.opened_by_session, 'alice-tmux-1');
-    }
-  } finally {
-    cleanup();
-  }
-});
-
-test('gate boot: parallel_session_authors flags actor with two sessions on same target (#249 slice 4)', () => {
-  const { root, cleanup } = bootstrap();
-  try {
-    // Same author (alice), same target, two distinct sessions — the
-    // self-race shape the slice exists to surface.
-    makeRequestSessioned(root, 'alice', 'work A', 'src/foo', 'alice-tmux-1');
-    makeRequestSessioned(root, 'alice', 'work B', 'src/foo', 'alice-tmux-2');
-
-    const { stdout } = runGate(root, ['boot']);
-    const payload = JSON.parse(stdout);
-    const entry = payload.active_overlapping_targets[0];
-    assert.deepEqual(entry.parallel_session_authors, {
-      alice: ['alice-tmux-1', 'alice-tmux-2'],
-    });
-  } finally {
-    cleanup();
-  }
-});
-
-test('gate boot: parallel_session_authors omitted when sessions match (#249 slice 4)', () => {
-  // Same author, same target, IDENTICAL session — that's not a
-  // self-race (the actor knows about both records). The map must
-  // stay absent so the "warn-on-divergence" contract holds.
-  const { root, cleanup } = bootstrap();
-  try {
-    makeRequestSessioned(root, 'alice', 'work A', 'src/foo', 'alice-tmux-1');
-    makeRequestSessioned(root, 'alice', 'work B', 'src/foo', 'alice-tmux-1');
-
-    const { stdout } = runGate(root, ['boot']);
-    const payload = JSON.parse(stdout);
-    const entry = payload.active_overlapping_targets[0];
-    assert.equal(
-      'parallel_session_authors' in entry,
-      false,
-      'identical sessions are not a self-race',
-    );
-  } finally {
-    cleanup();
-  }
-});
-
-test('gate boot: parallel_session_authors omitted when only one record stamped (#249 slice 4)', () => {
-  // One stamped, one unstamped — provenance unknown for the second
-  // record. Detection requires ≥2 of the actor's records to BOTH
-  // carry sessions AND those values to diverge. Mixed pre/post-#249
-  // groups should not falsely flag.
-  const { root, cleanup } = bootstrap();
-  try {
-    makeRequestSessioned(root, 'alice', 'work A', 'src/foo', 'alice-tmux-1');
-    makeRequestWithTarget(root, 'alice', 'work B', 'src/foo'); // no GUILD_SESSION_ID
-
-    const { stdout } = runGate(root, ['boot']);
-    const payload = JSON.parse(stdout);
-    const entry = payload.active_overlapping_targets[0];
-    assert.equal(
-      'parallel_session_authors' in entry,
-      false,
-      'one stamped + one unstamped is not enough to prove divergence',
-    );
-  } finally {
-    cleanup();
-  }
-});
-
-test('gate boot: different actors with different sessions do NOT flag (#249 slice 4)', () => {
-  // alice from session A, leysia from session B → genuine cross-
-  // session overlap (the original #234 surface), but NOT a
-  // SAME-actor parallel-session race. Only flag when a single
-  // actor's authorship splits.
-  const { root, cleanup } = bootstrap();
-  try {
-    registerMember(root, 'leysia');
-    makeRequestSessioned(root, 'alice', 'work A', 'src/foo', 'alice-tmux-1');
-    makeRequestSessioned(root, 'leysia', 'work B', 'src/foo', 'leysia-tmux-1');
-
-    const { stdout } = runGate(root, ['boot']);
-    const payload = JSON.parse(stdout);
-    const entry = payload.active_overlapping_targets[0];
-    assert.equal(
-      'parallel_session_authors' in entry,
-      false,
-      'different actors authoring is not a self-race',
-    );
-  } finally {
-    cleanup();
-  }
-});
-
-test('gate boot text: parallel-session warning lands under overlap section (#249 slice 4)', () => {
-  const { root, cleanup } = bootstrap();
-  try {
-    // Overlap text-mode rendering (and the parallel-session warn it
-    // hangs off) is profile=swarm only (#323).
-    writeFileSync(
-      join(root, 'guild.config.yaml'),
-      'content_root: .\nhost_names: [human]\nprofile: swarm\n',
-    );
-    makeRequestSessioned(root, 'alice', 'work A', 'src/foo', 'alice-tmux-1');
-    makeRequestSessioned(root, 'alice', 'work B', 'src/foo', 'alice-tmux-2');
-
-    const t = runGate(root, ['boot', '--format', 'text']);
-    assert.equal(t.status, 0);
-    assert.match(t.stdout, /active waves with overlapping target:/);
-    // The per-request line carries the bracket-tagged session.
-    assert.match(t.stdout, /\[session=alice-tmux-1\]/);
-    assert.match(t.stdout, /\[session=alice-tmux-2\]/);
-    // The dedicated warning line names the actor and both sessions.
-    assert.match(
-      t.stdout,
-      /⚠ same-actor parallel sessions: alice on target "src\/foo" \(sessions: alice-tmux-1, alice-tmux-2\)/,
-    );
-  } finally {
-    cleanup();
-  }
-});
-
-// -------------------- C3 / silent-fallback-loses-signal --------------------
-//
-// Combo C3 from the 2026-05-10 dogfood arc (substrate/agora/plays/
-// eris-dogfood-0510/2026-05-10-001.yaml). Pre-this-PR, four
-// enrichment paths in boot caught their own errors silently and the
-// payload's status counts went on lying. Devil's concern2 on PR #105
-// (agent-first-session 2026-04-16) flagged this as the gap that lets
-// agents trust the snapshot more than they should.
-
-test('#C3: clean boot has empty warnings array (no false positives)', () => {
-  const { root, cleanup } = bootstrap();
-  try {
-    const { stdout } = runGate(root, ['boot']);
-    const payload = JSON.parse(stdout);
-    assert.ok(Array.isArray(payload.warnings),
-      'warnings must always be an array (empty in clean case)');
-    assert.equal(payload.warnings.length, 0,
-      `clean bootstrap should produce zero warnings, got: ${JSON.stringify(payload.warnings)}`);
-
-    // text format must NOT print the warning block when empty
-    const { stdout: textOut } = runGate(root, ['boot', '--format', 'text']);
-    assert.doesNotMatch(textOut, /warning\(s\) raised/);
-  } finally {
-    cleanup();
-  }
-});
-
-test('#C3: broken issues path surfaces a warning instead of silent swallow', () => {
-  // Sabotage the issues directory: replace it with a regular file so
-  // listAll() throws when it tries to readdir(). This reproduces the
-  // exact silent-swallow path the original `// issues dir may not
-  // exist — non-fatal` catch was hiding.
-  const { root, cleanup } = bootstrap();
-  try {
-    // The issues path defaults to <content_root>/issues. Create a
-    // regular file at that path so the directory walk fails.
-    writeFileSync(join(root, 'issues'), 'NOT A DIRECTORY');
-    const { stdout, status } = runGate(root, ['boot']);
-    assert.equal(status, 0, 'boot must remain non-fatal even when enrichment fails');
-    const payload = JSON.parse(stdout);
-    assert.ok(Array.isArray(payload.warnings));
-    assert.ok(
-      payload.warnings.some((w: string) => /issues enrichment failed/.test(w)),
-      `expected an issues-enrichment warning; got: ${JSON.stringify(payload.warnings)}`,
-    );
-
-    // text format surfaces the warning block + the inaccuracy disclaimer
-    const { stdout: textOut } = runGate(root, ['boot', '--format', 'text']);
-    assert.match(textOut, /warning\(s\) raised/);
-    assert.match(textOut, /issues enrichment failed/);
-    assert.match(textOut, /counts in 'queues:'.*may be inaccurate/);
-  } finally {
-    cleanup();
-  }
-});
-
-test('#C3: warning text in JSON includes the underlying error message', () => {
-  // Silent try/catch dropped the error info entirely. The new shape
-  // must propagate enough of `e.message` for an agent reading the
-  // payload to know what actually broke.
-  const { root, cleanup } = bootstrap();
-  try {
-    writeFileSync(join(root, 'issues'), 'NOT A DIRECTORY');
-    const { stdout } = runGate(root, ['boot']);
-    const payload = JSON.parse(stdout);
-    const w = payload.warnings.find((s: string) =>
-      /issues enrichment failed/.test(s));
-    assert.ok(w, 'expected to find an issues-enrichment warning');
-    // The exact ENOTDIR / EEXIST shape varies by node version; just
-    // confirm SOMETHING about the error was preserved (not just the
-    // generic "issues enrichment failed").
-    assert.ok(
-      w.length > 'issues enrichment failed (open_issues count may be inaccurate): '.length + 5,
-      `warning should include the underlying error detail, got: ${w}`,
-    );
-  } finally {
-    cleanup();
-  }
-});
